Addressing comments from Vinoth
