+++ hudi-common/src/main/java/org/apache/hudi/common/bloom/filter/SimpleBloomFilter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.common;
+package org.apache.hudi.common.bloom.filter;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -25,48 +25,45 @@ import java.io.DataOutputStream;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import javax.xml.bind.DatatypeConverter;
-import org.apache.hadoop.util.bloom.DynamicBloomFilter;
 import org.apache.hadoop.util.bloom.Key;
 import org.apache.hudi.exception.HoodieIndexException;
 
 /**
- * Hudi's Dynamic Bloom Filter based out of {@link org.apache.hadoop.util.bloom.DynamicBloomFilter}
+ * A Simple Bloom filter implementation built on top of {@link org.apache.hadoop.util.bloom.BloomFilter}.
  */
-public class HoodieDynamicBloomFilter extends DynamicBloomFilter implements BloomFilter {
 
-  public static final String TYPE_CODE_PREFIX = "DYNAMIC";
-  public static final String TYPE_CODE = TYPE_CODE_PREFIX + "_V0";
-  private DynamicBloomFilter dynamicBloomFilter;
+public class SimpleBloomFilter implements BloomFilter {
+
+  public static final String TYPE_CODE = "SIMPLE";
+  private org.apache.hadoop.util.bloom.BloomFilter filter = null;
 
   /**
-   * Instantiates {@link HoodieDynamicBloomFilter} with the given args
+   * Create a new Bloom filter with the given configurations.
    *
    * @param numEntries The total number of entries.
    * @param errorRate maximum allowable error rate.
    * @param hashType type of the hashing function (see {@link org.apache.hadoop.util.hash.Hash}).
-   * @return the {@link HoodieDynamicBloomFilter} thus created
    */
-  HoodieDynamicBloomFilter(int numEntries, double errorRate, int hashType) {
+  public SimpleBloomFilter(int numEntries, double errorRate, int hashType) {
     // Bit size
     int bitSize = BloomFilterUtils.getBitSize(numEntries, errorRate);
     // Number of the hash functions
     int numHashs = BloomFilterUtils.getNumHashes(bitSize, numEntries);
-    this.dynamicBloomFilter = new DynamicBloomFilter(bitSize, numHashs, hashType, numEntries);
+    // The filter
+    this.filter = new org.apache.hadoop.util.bloom.BloomFilter(bitSize, numHashs, hashType);
   }
 
   /**
-   * Generate {@link HoodieDynamicBloomFilter} from the given {@code serString} serialized string
+   * Create the bloom filter from serialized string.
    *
-   * @param serString the serialized string which represents the {@link HoodieDynamicBloomFilter}
-   * @param typeCode type code of the bloom filter
+   * @param serString serialized string which represents the {@link SimpleBloomFilter}
    */
-  HoodieDynamicBloomFilter(String serString, String typeCode) {
-    // ignoring the type code for now, since we have just one version
+  public SimpleBloomFilter(String serString) {
+    this.filter = new org.apache.hadoop.util.bloom.BloomFilter();
     byte[] bytes = DatatypeConverter.parseBase64Binary(serString);
     DataInputStream dis = new DataInputStream(new ByteArrayInputStream(bytes));
     try {
-      dynamicBloomFilter = new DynamicBloomFilter();
-      dynamicBloomFilter.readFields(dis);
+      this.filter.readFields(dis);
       dis.close();
     } catch (IOException e) {
       throw new HoodieIndexException("Could not deserialize BloomFilter instance", e);
@@ -75,20 +72,29 @@ public class HoodieDynamicBloomFilter extends DynamicBloomFilter implements Bloo
 
   @Override
   public void add(String key) {
-    dynamicBloomFilter.add(new Key(key.getBytes(StandardCharsets.UTF_8)));
+    if (key == null) {
+      throw new NullPointerException("Key cannot by null");
+    }
+    filter.add(new Key(key.getBytes(StandardCharsets.UTF_8)));
   }
 
   @Override
   public boolean mightContain(String key) {
-    return dynamicBloomFilter.membershipTest(new Key(key.getBytes(StandardCharsets.UTF_8)));
+    if (key == null) {
+      throw new NullPointerException("Key cannot by null");
+    }
+    return filter.membershipTest(new Key(key.getBytes(StandardCharsets.UTF_8)));
   }
 
+  /**
+   * Serialize the bloom filter as a string.
+   */
   @Override
   public String serializeToString() {
     ByteArrayOutputStream baos = new ByteArrayOutputStream();
     DataOutputStream dos = new DataOutputStream(baos);
     try {
-      dynamicBloomFilter.write(dos);
+      filter.write(dos);
       byte[] bytes = baos.toByteArray();
       dos.close();
       return DatatypeConverter.printBase64Binary(bytes);
@@ -99,6 +105,6 @@ public class HoodieDynamicBloomFilter extends DynamicBloomFilter implements Bloo
 
   @Override
   public String getBloomFilterTypeCode() {
-    return HoodieDynamicBloomFilter.TYPE_CODE;
+    return SimpleBloomFilter.TYPE_CODE;
   }
 }
