+++ hudi-common/src/test/java/org/apache/hudi/common/TestBloomFilter.java
@@ -23,6 +23,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
+import org.apache.hudi.common.bloom.filter.BloomFilter;
+import org.apache.hudi.common.bloom.filter.BloomFilterFactory;
+import org.apache.hudi.common.bloom.filter.HoodieDynamicBloomFilter;
+import org.apache.hudi.common.bloom.filter.SimpleBloomFilter;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
